aws/credentials: Fix race condition with credentials expire
Fixes #834 by adding read write locks to calls to the credentials
provider's `IsExpired` method. Previously there was a race condition
where the credentials could be refreshing, set the Provider's Expiry,
while another goroutine is checking if the credentials are expired. In
this case the second goroutine would receive the old credentials,
incorrectly.

This change also adds protection to GetWithContext that the check for if
the credentials are not expired will not be blocked by the write mutex
in the provider's Retrieve method.

Incorporates some of the changes in
https://github.com/aws/aws-sdk-go/pull/3563 to remove the atomic value
around `creds`.