codegen: Dont serialize nil for visited nested event stream shapes
Updates the SDK's code gen to not encode entries as `nil` for shapes
that have already been visited. This skips `nil`'s in nested slice and
map event stream shapes.

Related to bigger issue of SDK's JSON serializer does not correctly
skip, or serailize NULL for nil values in slices.