Fix lexv1 AlternativeIntentsSerializationError
Deserialisation of the `X-Amz-Lex-Alternative-Intents` header fails because it
is a JSON array, but `aws.JSONValue` is a `map[string]interface{}`. Fix this by
treating `X-Amz-Lex-Alternative-Intents` as a base64 encoded string, which will
not be decoded and unmarshalled by the SDK. The caller of the SDK must manually
base64 decode and JSON unmarshal the field, in order to use the contents.