codegen: Add fix to prevent HTTP/2 event stream request hang on error (#4288)
Adds fix addressing an issue where SDK's bi-directional eventstream API
operation request could hang when the HTTP/2 stream was closed by the
service with a non-200 status code.

Go 1.15 through 1.17 have an issue where the HTTP/2 request can hang
while waiting for the HTTP response, and input payload reader. The
Expect: 100-Continue workaround forces the HTTP client to wait for a
`100 continue` or other HTTP status code before attempting to wait for
the request's body to be read. This workaround is not needed for Go
1.18, as a fix to the HTTP client resolves this issue.

Related to https://github.com/aws/aws-sdk-go-v2/pull/1515