aws/endpoints: Update unit tests use static data instead of live model (#4389)
Updates the SDK's endpoints model unit tests to ensure that unit tests
are performed on static data instead of the live model that can be
updated at any time. This change was done to ensure that upstream
sources of endpoint models making semantic changes do not cause unit
tests to break.