Update CI to include Go 1.19 (#4500)
Fixes issue with Go 1.19 discovered in the S3 transfer manager's windows
upload and downloader buffer pools. It looks like the update to Go 1.19
changes the code path that is used by the pool because io.NopCloser was
updated to forward WriterTo if the underlying Writer implemented it.
Looks likes we'd made some assumptions about this behavior not being
present. This PR suppresses the WriterTo of the buffer pool to prevent
the unexpected usage.