Add awsQueryCompatible trait support to service
When awsQueryCompatible trait is found, it's made available as an API property. When this property is found,
the error code is returned by looking up the mapping. This is a pre-requisite for migrating services from
AWSQuery wire protocol to AWSJson.

A test service has been generated for validating assumptions
Added unit tests as per the SEP