Initialise partition descriptors upon their first use.
The SDK used to initialise the descriptors of all partitions
statically. That has several downsides.

The first one is that descriptors of AWS partitions are always created
at application startup. That makes the startup slower, and wastes RAM
if the application only imports the SDK, but does not access Amazon
services. That is the case for tools like restic that link to clients
of multiple object storages. When working with GCS or Azure, it is
a waste of RAM to keep the descriptors of AWS partitions in RAM.

Instead, let us create the descriptors of AWS partitions when they
are first needed.

This turns out to be a profitable change for all applications. The go
compiler generates code differently when compiling functions for the
static init and when compiling normal functions that construct the
same structs. The static init produces some junk in .rodata section
of the compiled binary. On-demand init functions do not bring that
junk, and the binary becomes smaller by 1.5M.